Minor fixes
* Initialise members of `NammuController` in the constructor
* In `get_viewport_carets` test the type instead of catching all exceptions

The `AtfAreaView` issue was that sometimes the code "decided" that the pane was
split because of (presumably!) an exception from `top_ch, bottom_ch =
self.get_viewport_top_bottom(viewport)`.

The other was that some attributes of `NammuController` were not created in the
constructor, but only when opening a file. This only came up when running a
subset of the tests (the ones that call `arabic()`(?) without opening a file),
because when running the whole suite, a file has already been opened by previous
tests. I think it can't occur in normal usage, but it's cleaner to create the
attributes in the constructor anyway.